pyproject.toml
[tool.poetry]
name = "jaeger_mongodb_plugin"
version = "0.1.0"
description = ""
authors = ["Jeff Zambory <jeff.zambory@mongodb.com>"]
license = 'Apache-2.0'
packages = [
    { include = "jaeger_grpc_server", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.7"
protobuf = "^3.14.0"
google = "^3.0.0"
grpcio = "^1.35.0"
grpcio-health-checking = "^1.35.0"
pydantic = "^1.7.3"
pymongo = {extras = ["srv"], version = "^3.11.3"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-black = "^0.3.12"
pytest-cov = "^2.11.1"
pytest-flake8 = "^1.0.7"
pytest-mypy = "^0.8.0"
pytest-pydocstyle = "^2.2.0"
black = "^20.8b1"
ipython = "^7.19.0"
pytest-isort = "^1.3.0"

[tool.black]
line-length = 100
target-version = ['py37']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.pytest.ini_options]
flake8-ignore = "W605 W503 W291 E203 E501 F821"
addopts = "--flake8 --black --isort --pydocstyle"
testpaths = [
    "src"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
